refactor(rolldown_binding): extract plugin configs from binding_builtin_plugin
