+++ crates/rolldown_binding/src/options/plugin/config/binding_import_glob_plugin_config.rs
@@ -1,4 +1,4 @@
-use rolldown_plugin_import_glob::ImportGlobPluginConfig;
+use rolldown_plugin_import_glob::{ImportGlobPlugin, ImportGlobPluginConfig};
 
 #[napi_derive::napi(object)]
 #[derive(Debug, Default)]
@@ -7,11 +7,13 @@ pub struct BindingImportGlobPluginConfig {
   pub restore_query_extension: Option<bool>,
 }
 
-impl From<BindingImportGlobPluginConfig> for ImportGlobPluginConfig {
+impl From<BindingImportGlobPluginConfig> for ImportGlobPlugin {
   fn from(value: BindingImportGlobPluginConfig) -> Self {
-    ImportGlobPluginConfig {
-      root: value.root,
-      restore_query_extension: value.restore_query_extension.unwrap_or_default(),
+    Self {
+      config: ImportGlobPluginConfig {
+        root: value.root,
+        restore_query_extension: value.restore_query_extension.unwrap_or_default(),
+      },
     }
   }
 }
