+++ crates/rolldown_binding/src/options/plugin/config/binding_vite_resolve_plugin_config.rs
@@ -4,13 +4,11 @@ use std::sync::Arc;
 use napi::bindgen_prelude::FnArgs;
 use rolldown_plugin_vite_resolve::{
   FinalizeBareSpecifierCallback, FinalizeOtherSpecifiersCallback, ViteResolveOptions,
+  ViteResolveResolveOptions,
 };
 
 use crate::{
-  options::plugin::{
-    binding_builtin_plugin::BindingViteResolvePluginResolveOptions,
-    types::binding_limited_boolean::BindingTrueValue,
-  },
+  options::plugin::types::binding_limited_boolean::BindingTrueValue,
   types::{
     binding_string_or_regex::{BindingStringOrRegex, bindingify_string_or_regex_array},
     js_callback::{JsCallback, JsCallbackExt as _},
@@ -92,6 +90,49 @@ impl From<BindingViteResolvePluginConfig> for ViteResolveOptions {
           })
         },
       ),
+    }
+  }
+}
+
+#[napi_derive::napi(object)]
+#[derive(Debug)]
+#[allow(clippy::struct_excessive_bools)]
+pub struct BindingViteResolvePluginResolveOptions {
+  pub is_build: bool,
+  pub is_production: bool,
+  pub as_src: bool,
+  pub prefer_relative: bool,
+  pub is_require: Option<bool>,
+  pub root: String,
+  pub scan: bool,
+
+  pub main_fields: Vec<String>,
+  pub conditions: Vec<String>,
+  pub external_conditions: Vec<String>,
+  pub extensions: Vec<String>,
+  pub try_index: bool,
+  pub try_prefix: Option<String>,
+  pub preserve_symlinks: bool,
+}
+
+impl From<BindingViteResolvePluginResolveOptions> for ViteResolveResolveOptions {
+  fn from(value: BindingViteResolvePluginResolveOptions) -> Self {
+    Self {
+      is_build: value.is_build,
+      is_production: value.is_production,
+      as_src: value.as_src,
+      prefer_relative: value.prefer_relative,
+      is_require: value.is_require,
+      root: value.root,
+      scan: value.scan,
+
+      main_fields: value.main_fields,
+      conditions: value.conditions,
+      external_conditions: value.external_conditions,
+      extensions: value.extensions,
+      try_index: value.try_index,
+      try_prefix: value.try_prefix,
+      preserve_symlinks: value.preserve_symlinks,
     }
   }
 }
