+++ crates/rolldown_binding/src/options/plugin/config/mod.rs
@@ -1,7 +1,25 @@
+mod binding_alias_plugin_config;
+mod binding_asset_plugin_config;
+mod binding_build_import_analysis_plugin_config;
+mod binding_dynamic_import_vars_plugin_config;
 mod binding_import_glob_plugin_config;
+mod binding_isolated_declaration_plugin_config;
+mod binding_json_plugin_config;
+mod binding_manifest_plugin_config;
+mod binding_replace_plugin_config;
 mod binding_reporter_plugin_config;
+mod binding_transform_plugin_config;
 mod binding_vite_resolve_plugin_config;
 
+pub use binding_alias_plugin_config::BindingAliasPluginConfig;
+pub use binding_asset_plugin_config::BindingAssetPluginConfig;
+pub use binding_build_import_analysis_plugin_config::BindingBuildImportAnalysisPluginConfig;
+pub use binding_dynamic_import_vars_plugin_config::BindingDynamicImportVarsPluginConfig;
 pub use binding_import_glob_plugin_config::BindingImportGlobPluginConfig;
+pub use binding_isolated_declaration_plugin_config::BindingIsolatedDeclarationPluginConfig;
+pub use binding_json_plugin_config::BindingJsonPluginConfig;
+pub use binding_manifest_plugin_config::BindingManifestPluginConfig;
+pub use binding_replace_plugin_config::BindingReplacePluginConfig;
 pub use binding_reporter_plugin_config::BindingReporterPluginConfig;
+pub use binding_transform_plugin_config::BindingTransformPluginConfig;
 pub use binding_vite_resolve_plugin_config::BindingViteResolvePluginConfig;
